+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/BaseKtlintCheckTask.kt
@@ -5,7 +5,6 @@ import org.gradle.api.GradleException
 import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.FileCollection
 import org.gradle.api.file.FileTree
-import org.gradle.api.file.ProjectLayout
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
@@ -30,15 +29,14 @@ import java.io.PrintWriter
 
 @Suppress("UnstableApiUsage")
 abstract class BaseKtlintCheckTask(
-    private val objectFactory: ObjectFactory,
-    private val projectLayout: ProjectLayout
+    private val objectFactory: ObjectFactory
 ) : SourceTask() {
 
     @get:Classpath
     internal val classpath: ConfigurableFileCollection = project.files()
 
     @get:Internal
-    internal val additionalEditorconfigFile: RegularFileProperty = newFileProperty(objectFactory, projectLayout)
+    internal val additionalEditorconfigFile: RegularFileProperty = objectFactory.fileProperty()
 
     @get:PathSensitive(PathSensitivity.RELATIVE)
     @get:InputFiles
@@ -82,7 +80,7 @@ abstract class BaseKtlintCheckTask(
                     it.reporterName,
                     objectFactory.property { set(it.isAvailable()) },
                     it,
-                    newFileProperty(objectFactory, projectLayout).apply {
+                    objectFactory.fileProperty().apply {
                         set(it.getOutputFile())
                     }
                 )
@@ -95,7 +93,7 @@ abstract class BaseKtlintCheckTask(
             .map {
                 KtlintReport.CustomReport(
                     it.reporterId,
-                    newFileProperty(objectFactory, projectLayout).apply {
+                    objectFactory.fileProperty().apply {
                         set(
                             project.configurations.getByName(KTLINT_REPORTER_CONFIGURATION_NAME)
                                 .resolvedConfiguration
@@ -106,7 +104,7 @@ abstract class BaseKtlintCheckTask(
                                 ?.file ?: throw GradleException("Failed to resolve ${it.dependencyArtifact} artifact")
                         )
                     },
-                    newFileProperty(objectFactory, projectLayout).apply {
+                    objectFactory.fileProperty().apply {
                         set(it.getOutputFile())
                     }
                 )
@@ -142,7 +140,7 @@ abstract class BaseKtlintCheckTask(
     }
 
     @OutputFiles
-    private val ktlintArgsFile = newFileProperty(objectFactory, projectLayout).apply {
+    private val ktlintArgsFile = objectFactory.fileProperty().apply {
         set(
             project.layout.buildDirectory.file(
                 project.provider {