+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintCheckTask.kt
@@ -1,6 +1,5 @@
 package org.jlleitschuh.gradle.ktlint
 
-import org.gradle.api.file.ProjectLayout
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.tasks.CacheableTask
 import java.io.PrintWriter
@@ -9,8 +8,7 @@ import javax.inject.Inject
 @Suppress("UnstableApiUsage")
 @CacheableTask
 open class KtlintCheckTask @Inject constructor(
-    objectFactory: ObjectFactory,
-    projectLayout: ProjectLayout
-) : BaseKtlintCheckTask(objectFactory, projectLayout) {
+    objectFactory: ObjectFactory
+) : BaseKtlintCheckTask(objectFactory) {
     override fun additionalConfig(): (PrintWriter) -> Unit = {}
 }