+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintExtension.kt
@@ -4,7 +4,6 @@ import groovy.lang.Closure
 import org.gradle.api.Action
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.file.ConfigurableFileTree
-import org.gradle.api.file.ProjectLayout
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
@@ -22,7 +21,6 @@ import org.jlleitschuh.gradle.ktlint.reporter.ReporterType
 open class KtlintExtension
 internal constructor(
     objectFactory: ObjectFactory,
-    projectLayout: ProjectLayout,
     customReportersContainer: NamedDomainObjectContainer<CustomReporter>,
     private val filterTargetApplier: FilterApplier,
     kotlinScriptAdditionalPathApplier: KotlinScriptAdditionalPathApplier
@@ -90,7 +88,7 @@ internal constructor(
     /**
      * Provide additional `.editorconfig` file, that are not in the project or project parent folders.
      */
-    val additionalEditorconfigFile: RegularFileProperty = newFileProperty(objectFactory, projectLayout)
+    val additionalEditorconfigFile: RegularFileProperty = objectFactory.fileProperty()
 
     /**
      * Disable particular rules, by default enabled in ktlint, using rule id.